所有单据的布局样式调整，多订单下推全部调整

+++ app/src/main/java/com/fangzuo/assist/Activity/SettingActivity.java
@@ -1,13 +1,14 @@
 package com.fangzuo.assist.Activity;
 
 import android.app.AlertDialog;
-import android.app.ProgressDialog;
 import android.content.DialogInterface;
 import android.database.sqlite.SQLiteDatabase;
-import android.os.Bundle;
 import android.os.Handler;
 import android.os.Message;
 import android.support.design.widget.CoordinatorLayout;
+import android.support.v7.widget.DefaultItemAnimator;
+import android.support.v7.widget.DividerItemDecoration;
+import android.support.v7.widget.RecyclerView;
 import android.view.View;
 import android.widget.AdapterView;
 import android.widget.Button;
@@ -17,31 +18,27 @@ import android.widget.RelativeLayout;
 import android.widget.TextView;
 
 import com.fangzuo.assist.ABase.BaseActivity;
+import com.fangzuo.assist.Activity.Crash.App;
+import com.fangzuo.assist.Adapter.DataSearchRyAdapter;
 import com.fangzuo.assist.Beans.CommonResponse;
 import com.fangzuo.assist.Beans.ConnectResponseBean;
 import com.fangzuo.assist.Beans.EventBusEvent.ClassEvent;
 import com.fangzuo.assist.R;
-import com.fangzuo.assist.Server.WebAPI;
+import com.fangzuo.assist.RxSerivce.MySubscribe;
 import com.fangzuo.assist.Service.DataService;
-import com.fangzuo.assist.Utils.Asynchttp;
 import com.fangzuo.assist.Utils.BasicShareUtil;
-import com.fangzuo.assist.Utils.CallBack;
 import com.fangzuo.assist.Utils.DataBaseAdapter;
 import com.fangzuo.assist.Utils.DownLoadData;
 import com.fangzuo.assist.Utils.EventBusInfoCode;
 import com.fangzuo.assist.Utils.GreenDaoManager;
 import com.fangzuo.assist.Utils.Info;
 import com.fangzuo.assist.Utils.JsonCreater;
-import com.fangzuo.assist.Utils.RetrofitUtil;
 import com.fangzuo.assist.Utils.ShareUtil;
-import com.fangzuo.assist.Utils.SnackBarUtil;
 import com.fangzuo.assist.Utils.Toast;
-import com.fangzuo.assist.Utils.WebApi;
 import com.fangzuo.assist.widget.LoadingUtil;
 import com.fangzuo.greendao.gen.DaoMaster;
 import com.fangzuo.greendao.gen.DaoSession;
 import com.google.gson.Gson;
-import com.loopj.android.http.AsyncHttpClient;
 
 import java.util.ArrayList;
 
@@ -50,12 +47,14 @@ import butterknife.ButterKnife;
 import butterknife.OnClick;
 
 
-public class SettingActivity extends BaseActivity {
+public class SettingActivity extends BaseActivity  implements DataSearchRyAdapter.OnItemClickListener {
 
     @BindView(R.id.btn_back)
     RelativeLayout btnBack;
     @BindView(R.id.tv_title)
     TextView tvTitle;
+    @BindView(R.id.ry_data_search)
+    RecyclerView ryDataSearch;
     private DataBaseAdapter adapter;
     private ListView mLvDataBase;
     private EditText mEtUserName;
@@ -67,8 +66,7 @@ public class SettingActivity extends BaseActivity {
     private Button mBtnDownload;
     private SettingActivity mContext;
     private CommonListener commonListener;
-    private Gson gson;
-//    private ProgressDialog pg;
+    //    private ProgressDialog pg;
     private ArrayList<ConnectResponseBean.DataBaseList> container;
     private BasicShareUtil share;
     private String chooseDatabase;
@@ -76,7 +74,7 @@ public class SettingActivity extends BaseActivity {
     private SQLiteDatabase db;
     private DaoMaster daoMaster;
     private DaoSession daoSession;
-    private DaoSession session;
+    private DataSearchRyAdapter dataSearchRyAdapter;
 
 
     Handler handler = new Handler() {
@@ -160,7 +158,6 @@ public class SettingActivity extends BaseActivity {
     @Override
     public void initData() {
         mContext = this;
-        gson = new Gson();
         container = new ArrayList<>();
         commonListener = new CommonListener();
 //
@@ -177,7 +174,12 @@ public class SettingActivity extends BaseActivity {
             mEtPassword.setText(share.getDataBasePass());
         }
 
-        session = GreenDaoManager.getmInstance(mContext).getDaoSession();
+        dataSearchRyAdapter = new DataSearchRyAdapter(mContext, container);
+        ryDataSearch.setAdapter(dataSearchRyAdapter);
+        ryDataSearch.setItemAnimator(new DefaultItemAnimator());
+        ryDataSearch.addItemDecoration(new DividerItemDecoration(mContext, DividerItemDecoration.VERTICAL));
+        dataSearchRyAdapter.setOnItemClickListener(this);
+        dataSearchRyAdapter.notifyDataSetChanged();
     }
 
     @Override
@@ -185,6 +187,8 @@ public class SettingActivity extends BaseActivity {
         mBtnConn.setOnClickListener(commonListener);
         mBtnProp.setOnClickListener(commonListener);
         mBtnDownload.setOnClickListener(commonListener);
+
+
         mLvDataBase.setOnItemClickListener(new AdapterView.OnItemClickListener() {
             @Override
             public void onItemClick(AdapterView<?> adapterView, View view, int i, long l) {
@@ -215,6 +219,21 @@ public class SettingActivity extends BaseActivity {
             case R.id.tv_title:
                 break;
         }
+    }
+
+    @Override
+    public void onItemClick(View view, int position) {
+
+//        adapter.setIsCheck(position);
+//        adapter.notifyDataSetChanged();
+        dataSearchRyAdapter.setIsCheck(position);
+        chooseDatabase = container.get(position).dataBaseName;
+        Toast.showText(mContext, chooseDatabase);
+    }
+
+    @Override
+    public void onItemLongClick(View view, int position) {
+
     }
 
     private class CommonListener implements View.OnClickListener {
@@ -259,31 +278,54 @@ public class SettingActivity extends BaseActivity {
     }
 
     private void setprop(final boolean isClear) {
-        LoadingUtil.show(mContext,"正在配置...");
+        if (null==chooseDatabase){
+            Toast.showText(mContext,"请选择账套");
+            return;
+        }
+        LoadingUtil.show(mContext, "正在配置...");
         final AlertDialog.Builder ab = new AlertDialog.Builder(mContext);
-        String json = JsonCreater.ConnectSQL(share.getDatabaseIp(), share.getDatabasePort(), share.getDataBaseUser(), share.getDataBasePass(), chooseDatabase);
-        RetrofitUtil.getInstance(mContext).createReq(WebAPI.class).SetProp(RetrofitUtil.getParams(mContext, json))
-                .enqueue(new CallBack() {
+        App.getRService().SetProp(
+                JsonCreater.ConnectSQL(
+                        share.getDatabaseIp(),
+                        share.getDatabasePort(),
+                        share.getDataBaseUser(),
+                        share.getDataBasePass(),
+                        chooseDatabase), new MySubscribe<CommonResponse>() {
                     @Override
-                    public void onSucceed(CommonResponse cBean) {
-                        if (isClear) {
-                            DataService.deleteAll(mContext);
-                            deleteData();
+                    public void onNext(CommonResponse commonResponse) {
+                        if (commonResponse.state) {
+                            if (isClear) {
+                                DataService.deleteAll(mContext);
+                                deleteData();
+                            }
+                            LoadingUtil.dismiss();
+                            ab.setTitle("配置结果");
+                            ab.setMessage("配置成功，请继续下一步操作");
+                            ab.setPositiveButton("确认", null);
+                            ab.create().show();
+                            share.setVersion(commonResponse.returnJson);
+                            share.setDataBase(chooseDatabase);
+                        } else {
+                            LoadingUtil.dismiss();
+                            ab.setTitle("配置结果");
+                            ab.setMessage(commonResponse.returnJson);
+                            ab.setPositiveButton("确认", null);
+                            ab.setNegativeButton("重试", new DialogInterface.OnClickListener() {
+                                @Override
+                                public void onClick(DialogInterface dialogInterface, int i) {
+                                    prop();
+                                }
+                            });
+                            ab.create().show();
                         }
-                        LoadingUtil.dismiss();
-                        ab.setTitle("配置结果");
-                        ab.setMessage("配置成功，请继续下一步操作");
-                        ab.setPositiveButton("确认", null);
-                        ab.create().show();
-                        share.setVersion(cBean.returnJson);
-                        share.setDataBase(chooseDatabase);
+
                     }
 
                     @Override
-                    public void OnFail(String Msg) {
+                    public void onError(Throwable e) {
                         LoadingUtil.dismiss();
                         ab.setTitle("配置结果");
-                        ab.setMessage(Msg);
+                        ab.setMessage(e.toString());
                         ab.setPositiveButton("确认", null);
                         ab.setNegativeButton("重试", new DialogInterface.OnClickListener() {
                             @Override
@@ -294,6 +336,40 @@ public class SettingActivity extends BaseActivity {
                         ab.create().show();
                     }
                 });
+
+//        String json = JsonCreater.ConnectSQL(share.getDatabaseIp(), share.getDatabasePort(), share.getDataBaseUser(), share.getDataBasePass(), chooseDatabase);
+//        RetrofitUtil.getInstance(mContext).createReq(WebAPI.class).SetProp(RetrofitUtil.getParams(mContext, json))
+//                .enqueue(new CallBack() {
+//                    @Override
+//                    public void onSucceed(CommonResponse cBean) {
+//                        if (isClear) {
+//                            DataService.deleteAll(mContext);
+//                            deleteData();
+//                        }
+//                        LoadingUtil.dismiss();
+//                        ab.setTitle("配置结果");
+//                        ab.setMessage("配置成功，请继续下一步操作");
+//                        ab.setPositiveButton("确认", null);
+//                        ab.create().show();
+//                        share.setVersion(cBean.returnJson);
+//                        share.setDataBase(chooseDatabase);
+//                    }
+//
+//                    @Override
+//                    public void OnFail(String Msg) {
+//                        LoadingUtil.dismiss();
+//                        ab.setTitle("配置结果");
+//                        ab.setMessage(Msg);
+//                        ab.setPositiveButton("确认", null);
+//                        ab.setNegativeButton("重试", new DialogInterface.OnClickListener() {
+//                            @Override
+//                            public void onClick(DialogInterface dialogInterface, int i) {
+//                                prop();
+//                            }
+//                        });
+//                        ab.create().show();
+//                    }
+//                });
     }
 
     private void deleteData() {
@@ -329,46 +405,82 @@ public class SettingActivity extends BaseActivity {
     }
 
     private void connectToSQL() {
-        LoadingUtil.show(mContext,"正在连接...");
-        String json = JsonCreater.ConnectSQL(mEtServerIP.getText().toString(), mEtServerPort.getText().toString(),
-                mEtUserName.getText().toString(), mEtPassword.getText().toString(), Info.DATABASESETTING);
-        Asynchttp.post(mContext, getBaseUrl() + WebApi.CONNECTSQL, json, new Asynchttp.Response() {
+        LoadingUtil.showDialog(mContext, "正在连接...");
+        App.getRService().connectToSQL(
+                JsonCreater.ConnectSQL(
+                        mEtServerIP.getText().toString(),
+                        mEtServerPort.getText().toString(),
+                        mEtUserName.getText().toString(),
+                        mEtPassword.getText().toString(),
+                        Info.DATABASESETTING), new MySubscribe<CommonResponse>() {
+                    @Override
+                    public void onNext(CommonResponse commonResponse) {
+                        share.setDatabaseIp(mEtServerIP.getText().toString());
+                        share.setDatabasePort(mEtServerPort.getText().toString());
+                        share.setDataBaseUser(mEtUserName.getText().toString());
+                        share.setDataBasePass(mEtPassword.getText().toString());
+                        LoadingUtil.dismiss();
+                        ConnectResponseBean connectBean = gson.fromJson(commonResponse.returnJson, ConnectResponseBean.class);
+//                        container.clear();
+//                        ConnectResponseBean connectResponseBean = new ConnectResponseBean();
+//                        ConnectResponseBean.DataBaseList dBean = connectResponseBean.new DataBaseList();
+//                        dBean.name = "账套";
+//                        dBean.dataBaseName = "数据库";
+//                        container.add(dBean);
+//                        container.addAll(connectBean.DataBaseList);
+                        dataSearchRyAdapter.addAll(connectBean.DataBaseList);
 
+                        adapter = new DataBaseAdapter(mContext, container);
+                        mLvDataBase.setAdapter(adapter);
+                        Toast.showText(mContext, "获取了" + connectBean.DataBaseList.size() + "条数据");
+                    }
 
-            @Override
-            public void onSucceed(CommonResponse cBean, AsyncHttpClient client) {
-                share.setDatabaseIp(mEtServerIP.getText().toString());
-                share.setDatabasePort(mEtServerPort.getText().toString());
-                share.setDataBaseUser(mEtUserName.getText().toString());
-                share.setDataBasePass(mEtPassword.getText().toString());
-                LoadingUtil.dismiss();
-                ConnectResponseBean connectBean = gson.fromJson(cBean.returnJson, ConnectResponseBean.class);
-                container.clear();
-                ConnectResponseBean connectResponseBean = new ConnectResponseBean();
-                ConnectResponseBean.DataBaseList dBean = connectResponseBean.new DataBaseList();
-                dBean.name = "账套";
-                dBean.dataBaseName = "数据库";
-                container.add(dBean);
-                container.addAll(connectBean.DataBaseList);
-                adapter = new DataBaseAdapter(mContext, container);
-                mLvDataBase.setAdapter(adapter);
-                Toast.showText(mContext, "获取了" + connectBean.DataBaseList.size() + "条数据");
-            }
+                    @Override
+                    public void onError(Throwable e) {
+                        LoadingUtil.dismiss();
+                        Toast.showText(mContext, "连接错误");
+                    }
+                });
 
-            @Override
-            public void onFailed(String Msg, AsyncHttpClient client) {
-                LoadingUtil.dismiss();
-                Toast.showText(mContext,"连接失败，请重试");
-//                SnackBarUtil.LongSnackbar(containerView, Msg, SnackBarUtil.Alert).setAction("重试", new View.OnClickListener() {
-//                    @Override
-//                    public void onClick(View view) {
-//                        connectToSQL();
-//                    }
-//                }).show();
-
-
-            }
-        });
+//        String json = JsonCreater.ConnectSQL(mEtServerIP.getText().toString(), mEtServerPort.getText().toString(),
+//                mEtUserName.getText().toString(), mEtPassword.getText().toString(), Info.DATABASESETTING);
+//        Asynchttp.post(mContext, getBaseUrl() + WebApi.CONNECTSQL, json, new Asynchttp.Response() {
+//
+//
+//            @Override
+//            public void onSucceed(CommonResponse cBean, AsyncHttpClient client) {
+//                share.setDatabaseIp(mEtServerIP.getText().toString());
+//                share.setDatabasePort(mEtServerPort.getText().toString());
+//                share.setDataBaseUser(mEtUserName.getText().toString());
+//                share.setDataBasePass(mEtPassword.getText().toString());
+//                LoadingUtil.dismiss();
+//                ConnectResponseBean connectBean = gson.fromJson(cBean.returnJson, ConnectResponseBean.class);
+//                container.clear();
+//                ConnectResponseBean connectResponseBean = new ConnectResponseBean();
+//                ConnectResponseBean.DataBaseList dBean = connectResponseBean.new DataBaseList();
+//                dBean.name = "账套";
+//                dBean.dataBaseName = "数据库";
+//                container.add(dBean);
+//                container.addAll(connectBean.DataBaseList);
+//                adapter = new DataBaseAdapter(mContext, container);
+//                mLvDataBase.setAdapter(adapter);
+//                Toast.showText(mContext, "获取了" + connectBean.DataBaseList.size() + "条数据");
+//            }
+//
+//            @Override
+//            public void onFailed(String Msg, AsyncHttpClient client) {
+//                LoadingUtil.dismiss();
+//                Toast.showText(mContext,"连接失败，请重试");
+////                SnackBarUtil.LongSnackbar(containerView, Msg, SnackBarUtil.Alert).setAction("重试", new View.OnClickListener() {
+////                    @Override
+////                    public void onClick(View view) {
+////                        connectToSQL();
+////                    }
+////                }).show();
+//
+//
+//            }
+//        });
     }
